block APIs outside of development

+++ packages/frontend/src/pages/api/delete-all.ts
@@ -5,6 +5,11 @@ export default async function deleteAll(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
+  if (process.env.NODE_ENV !== 'development') {
+    res.status(404).end();
+    return;
+  }
+
   await deleteAllData();
   res.status(200).json({ deleted: true });
 }
